requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
braintree==4.19.0
Brotli==1.0.9
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cssselect2==0.7.0
Django==4.1.2
django-braces==1.15.0
django-model-utils==4.3.1
django-star-ratings==0.9.2
fonttools==4.39.4
html5lib==1.1
idna==3.4
kombu==5.2.4
Pillow==9.5.0
prompt-toolkit==3.0.38
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
pytz==2023.3
redis==4.5.5
requests==2.30.0
six==1.16.0
sqlparse==0.4.4
swapper==1.3.0
tinycss2==1.2.1
tzdata==2023.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
weasyprint==59.0
webencodings==0.5.1
zopfli==0.2.2
